Now I'm abble to send little portions of captured data( sound and vibration values ) over wifi connection. Tested with 16Hz...

+++ include/fallsensor.h
@@ -14,10 +14,10 @@
 
 typedef struct{
 	struct{
-		unsigned long _0[AMOST_BUFFER_SIZE];
-		unsigned long _1[AMOST_BUFFER_SIZE];
-		unsigned long *active;
-		unsigned long *toSave;
+		unsigned short _0[AMOST_BUFFER_SIZE];
+		unsigned short _1[AMOST_BUFFER_SIZE];
+		unsigned short *active;
+		unsigned short *toSave;
 		unsigned int size;
 	}buff;
 	struct{
